+++ .devcontainer/codespaces_init.sh
@@ -3,3 +3,4 @@
 (echo && cat ./.devcontainer/aliases.sh) >> ~/.bashrc
 
 git config --global push.autoSetupRemote true
+cp ./.devcontainer/pre-commit ./.git/hooks/pre-commit