+++ infinite_games/utils/config.py
@@ -62,7 +62,8 @@ def add_args(cls, parser):
     Adds relevant arguments to the parser for operation.
     """
     # Netuid Arg: The netuid of the subnet to connect to.
-    parser.add_argument("--netuid", type=int, help="Subnet netuid", default=os.getenv("NETUID", 1))
+    # defaults to 6 (the mainnet IG).
+    parser.add_argument("--netuid", type=int, help="Subnet netuid", default=os.getenv("NETUID", 6))
 
     neuron_type = "validator" if "miner" not in cls.__name__.lower() else "miner"