+++ infinite_games/utils/misc.py
@@ -97,8 +97,8 @@ def _ttl_hash_gen(seconds: int):
 @backoff.on_exception(
     backoff.constant,
     Exception,
-    interval=0.5,
-    max_time=10,
+    interval=1,
+    max_time=20,
     max_tries=10,
     on_backoff=lambda details: bt.logging.warning(
         f"Retrying get block due to exception: {details['exception']}"
@@ -127,18 +127,7 @@ def ttl_get_block(self) -> int:
     Note: self here is the miner or validator instance
     """
 
-    # get_current_block often errors in testnet, so we override it here
-    if self.subtensor.network in ["test", "mock"]:
-        # seen in logs
-        # 2024-11-26 18:21:05.770 |  Validator starting at block: 3322388
-        start_ts_str = "2024-11-26 18:21:05"
-        start_ts = datetime.strptime(start_ts_str, "%Y-%m-%d %H:%M:%S")
-        time_diff = datetime.now() - start_ts
-        n_blocks = int(time_diff.total_seconds() / 12)
-        start_block = 3322388
-        return start_block + n_blocks
-    else:
-        return self.subtensor.get_current_block()
+    return self.subtensor.get_current_block()
 
 
 async def split_chunks(l, n):